the last line stores in numberOfObjects

+++ src/main/java/edu/ucr/cs236/RankSorting.java
@@ -30,7 +30,7 @@ public class RankSorting {
 		job.setMapOutputKeyClass(Text.class);
 		job.setMapOutputValueClass(Text.class);
 
-		job.setOutputKeyClass(IntWritable.class);
+		job.setOutputKeyClass(LongWritable.class);
 		job.setOutputValueClass(Text.class);
 
 		job.setMapperClass(RankSortingMapper.class);
@@ -62,12 +62,12 @@ public class RankSorting {
 
 	}
 
-	public static class RankSortingReducer extends Reducer<Text, Text, IntWritable, Text> {
+	public static class RankSortingReducer extends Reducer<Text, Text, LongWritable, Text> {
 		@Override
 		protected void reduce(Text key, java.lang.Iterable<Text> values, Context context) throws IOException, InterruptedException {
 			int i = 1;
 			for (Text t : values){
-				context.write(new IntWritable(i),new Text(key.toString().substring(0, key.toString().indexOf(":") + 1) + t));//new Text(key.toString().substring(0, key.find(":"))), t);
+				context.write(new LongWritable(i),new Text(key.toString().substring(0, key.toString().indexOf(":") + 1) + t));//new Text(key.toString().substring(0, key.find(":"))), t);
 				i++;
 			}
 		}
